+++ packages/astro/test/fixtures/astro-markdown-shiki/langs/src/layouts/content.astro
