+++ packages/astro/test/fixtures/astro-markdown-shiki/normal/astro.config.mjs
@@ -1,0 +1,8 @@
+export default {
+	markdownOptions: {
+		render: ['@astrojs/markdown-remark', { syntaxHighlight: 'shiki' }],
+	},
+	buildOptions: {
+		sitemap: false,
+	},
+}
