+++ packages/astro/test/fixtures/astro-markdown-shiki/normal/src/layouts/content.astro
