+++ packages/astro/test/fixtures/astro-markdown-shiki/normal/src/pages/astro.astro
