+++ packages/astro/test/fixtures/astro-markdown-shiki/themes-custom/src/layouts/content.astro
