+++ packages/astro/test/fixtures/astro-markdown-shiki/themes-integrated/astro.config.mjs
@@ -1,0 +1,8 @@
+export default {
+	markdownOptions: {
+		render: ["@astrojs/markdown-remark", { syntaxHighlight: 'shiki', shikiConfig: { theme: 'github-light' } }],
+	},
+	buildOptions: {
+		sitemap: false,
+	},
+}
