+++ packages/astro/test/fixtures/astro-markdown-shiki/themes-integrated/src/pages/astro.astro
@@ -1,0 +1,14 @@
+---
+import { Markdown } from 'astro/components';
+import Layout from '../layouts/content.astro';
+---
+
+<Layout>
+	<Markdown>
+		# Hello world
+
+		```js
+			console.log('JavaScript')
+		```
+	</Markdown>
+</Layout>
