+++ packages/astro/test/fixtures/astro-markdown-shiki/wrap-false/astro.config.mjs
@@ -1,0 +1,8 @@
+export default {
+	markdownOptions: {
+		render: ['@astrojs/markdown-remark', { syntaxHighlight: 'shiki', shikiConfig: { wrap: false } }],
+	},
+	buildOptions: {
+		sitemap: false,
+	},
+}
