+++ packages/astro/test/fixtures/astro-markdown-shiki/wrap-false/src/pages/astro.astro
@@ -1,0 +1,18 @@
+---
+import { Markdown } from 'astro/components';
+import Layout from '../layouts/content.astro';
+---
+
+<Layout>
+	<Markdown>
+		# Hello world
+
+		```
+			plaintext
+		```
+
+		```js
+			console.log('JavaScript')
+		```
+	</Markdown>
+</Layout>
