+++ packages/astro/test/fixtures/astro-markdown-shiki/wrap-null/astro.config.mjs
@@ -1,0 +1,8 @@
+export default {
+	markdownOptions: {
+		render: ['@astrojs/markdown-remark', { syntaxHighlight: 'shiki', shikiConfig: { wrap: null } }],
+	},
+	buildOptions: {
+		sitemap: false,
+	},
+}
