+++ packages/astro/test/fixtures/astro-markdown-shiki/wrap-null/src/layouts/content.astro
@@ -1,0 +1,10 @@
+<html>
+  <head>
+    <!-- Head Stuff -->
+  </head>
+  <body>
+    <div class="container">
+      <slot></slot>
+    </div>
+  </body>
+</html>
