+++ packages/astro/test/fixtures/astro-markdown-shiki/wrap-true/astro.config.mjs
@@ -1,0 +1,8 @@
+export default {
+	markdownOptions: {
+		render: ['@astrojs/markdown-remark', { syntaxHighlight: 'shiki', shikiConfig: { wrap: true } }],
+	},
+	buildOptions: {
+		sitemap: false,
+	},
+}
